move dto to api side + start conservation project param validation
